Separate response generator class and make it non-static
